move credentialDisplayMap to separate file
